// src/models/ITimelineActivity.ts
// Represents attributes of timeline activity
export interface ITimelineActivity {
    id: number;
    activityTitle: string;
    acivitySDate: Date;
    acivityEDate: Date;
    activityPilote: string;
    activityDirection: string;
    activityDescription: string;
    activityActions: string;
    activityLink: string;
}